Work in progress FilterableWorkflowTable
