+++ src/workflows-app/SubmissionDetails.js
@@ -1,21 +1,16 @@
 import _ from 'lodash/fp';
 import { useCallback, useMemo, useRef, useState } from 'react';
-import { div, h, h2, h3, span } from 'react-hyperscript-helpers';
-import { AutoSizer } from 'react-virtualized';
-import { ClipboardButton } from 'src/components/ClipboardButton';
-import { ButtonPrimary, Link, Select } from 'src/components/common';
+import { div, h, h2, h3 } from 'react-hyperscript-helpers';
+import { Link } from 'src/components/common';
 import { centeredSpinner, icon } from 'src/components/icons';
-import Modal from 'src/components/Modal';
-import { FlexTable, paginator, Sortable, tableHeight, TextCell } from 'src/components/table';
 import { Ajax } from 'src/libs/ajax';
-import colors from 'src/libs/colors';
 import * as Nav from 'src/libs/nav';
 import { notify } from 'src/libs/notifications';
 import { useCancellation, useOnMount, usePollingEffect } from 'src/libs/react-utils';
 import { AppProxyUrlStatus, workflowsAppStore } from 'src/libs/state';
 import * as Utils from 'src/libs/utils';
-import { customFormatDuration, differenceFromNowInSeconds, makeCompleteDate } from 'src/libs/utils';
-import { HeaderSection, statusType, SubmitNewWorkflowButton } from 'src/workflows-app/components/job-common';
+import { customFormatDuration, makeCompleteDate } from 'src/libs/utils';
+import { HeaderSection, SubmitNewWorkflowButton } from 'src/workflows-app/components/job-common';
 import { doesAppProxyUrlExist, loadAppUrls } from 'src/workflows-app/utils/app-utils';
 import {
   AutoRefreshInterval,
@@ -23,68 +18,23 @@ import {
   getDuration,
   isRunInTerminalState,
   isRunSetInTerminalState,
-  makeStatusLine,
 } from 'src/workflows-app/utils/submission-utils';
 import { wrapWorkflowsPage } from 'src/workflows-app/WorkflowsContainer';
 
+import FilterableWorkflowTable from './components/FilterableWorkflowTable';
+
 export const BaseSubmissionDetails = ({ name, namespace, workspace, submissionId }, _ref) => {
   // State
-  const [sort, setSort] = useState({ field: 'duration', direction: 'desc' });
-  const [pageNumber, setPageNumber] = useState(1);
-  const [itemsPerPage, setItemsPerPage] = useState(50);
-  const [viewErrorsId, setViewErrorsId] = useState();
   const [runsData, setRunsData] = useState();
   const [runsFullyUpdated, setRunsFullyUpdated] = useState();
   const [loading, setLoading] = useState(false);
 
   const [runSetData, setRunSetData] = useState();
   const [methodsData, setMethodsData] = useState();
-  const [filterOption, setFilterOption] = useState(null);
 
   const signal = useCancellation();
   const scheduledRefresh = useRef();
   const workspaceId = workspace.workspace.workspaceId;
-
-  const getFilter = (filterOption) => {
-    let filterStatement;
-    switch (filterOption) {
-      case 'Error':
-        filterStatement = _.filter((r) => errorStates.includes(r.state));
-        break;
-      case 'Succeeded':
-        filterStatement = _.filter((r) => r.state === 'COMPLETE');
-        break;
-      default:
-        filterStatement = (data) => data;
-    }
-    return filterStatement;
-  };
-
-  const state = (state, submissionDate) => {
-    switch (state) {
-      case 'SYSTEM_ERROR':
-      case 'EXECUTOR_ERROR':
-        return statusType.failed;
-      case 'COMPLETE':
-        return statusType.succeeded;
-      case 'INITIALIZING':
-        return statusType.initializing;
-      case 'QUEUED':
-        return statusType.queued;
-      case 'RUNNING':
-        return statusType.running;
-      case 'PAUSED':
-        return statusType.paused;
-      case 'CANCELED':
-        return statusType.canceled;
-      case 'CANCELING':
-        return statusType.canceling;
-      default:
-        // 10 seconds should be enough for Cromwell to summarize the new workflow and get a status other
-        // than UNKNOWN. In the meantime, handle this as an edge case in the UI:
-        return differenceFromNowInSeconds(submissionDate) < 10 ? statusType.initializing : statusType.unknown;
-    }
-  };
 
   const loadMethodsData = useCallback(
     async (cbasUrlRoot, methodId, methodVersion) => {
@@ -220,15 +170,6 @@ export const BaseSubmissionDetails = ({ name, namespace, workspace, submissionId
   const methodId = filteredRunSets[0]?.method_id;
   const getSpecificMethod = _.filter((m) => m.method_id === methodId, methodsData);
 
-  const errorStates = ['SYSTEM_ERROR', 'EXECUTOR_ERROR'];
-  const filteredPreviousRuns = filterOption ? getFilter(filterOption)(runsData) : runsData;
-  const sortedPreviousRuns = _.orderBy(sort.field, sort.direction, filteredPreviousRuns);
-  const filterOptions = ['Error', 'No filter', 'Succeeded'];
-
-  const firstPageIndex = (pageNumber - 1) * itemsPerPage;
-  const lastPageIndex = firstPageIndex + itemsPerPage;
-  const paginatedPreviousRuns = sortedPreviousRuns.slice(firstPageIndex, lastPageIndex);
-
   const header = useMemo(() => {
     const breadcrumbPathObjects = [
       {
@@ -244,8 +185,6 @@ export const BaseSubmissionDetails = ({ name, namespace, workspace, submissionId
     return h(HeaderSection, { breadcrumbPathObjects, button: h(SubmitNewWorkflowButton, { name, namespace }), title: 'Submission Details' });
   }, [name, namespace, submissionId]);
 
-  const rowWidth = 100;
-  const rowHeight = 50;
   return loading
     ? centeredSpinner()
     : div({ id: 'submission-details-page' }, [
@@ -298,179 +237,7 @@ export const BaseSubmissionDetails = ({ name, namespace, workspace, submissionId
             ]),
           ]
         ),
-        div(
-          {
-            style: {
-              backgroundColor: 'rgb(235, 236, 238)',
-              display: 'flex',
-              flex: '1 1 auto',
-              flexDirection: 'column',
-              padding: '1rem 3rem',
-            },
-          },
-          [
-            div(
-              {
-                style: {
-                  marginTop: '1em',
-                  height: tableHeight({ actualRows: paginatedPreviousRuns.length, maxRows: 12.5, heightPerRow: 250 }),
-                  minHeight: '10em',
-                },
-              },
-              [
-                div([h2(['Workflows'])]),
-                runsFullyUpdated
-                  ? div([icon('check', { size: 15, style: { color: colors.success() } }), ' Workflow statuses are all up to date.'])
-                  : div([
-                      icon('warning-standard', { size: 15, style: { color: colors.warning() } }),
-                      ' Some workflow statuses are not up to date. Refreshing the page may update more statuses.',
-                    ]),
-                div([h3(['Filter by: '])]),
-                h(Select, {
-                  isDisabled: false,
-                  'aria-label': 'Filter selection',
-                  isClearable: false,
-                  value: filterOption,
-                  placeholder: 'None selected',
-                  onChange: ({ value }) => {
-                    setFilterOption(value);
-                  },
-                  styles: { container: (old) => ({ ...old, display: 'inline-block', width: 200, marginBottom: '1.5rem' }) },
-                  options: filterOptions,
-                }),
-                h(AutoSizer, [
-                  ({ width, height }) =>
-                    h(FlexTable, {
-                      'aria-label': 'previous runs',
-                      width,
-                      height,
-                      sort,
-                      rowCount: paginatedPreviousRuns.length,
-                      noContentMessage: 'Nothing here yet! Your previously run workflows will be displayed here.',
-                      hoverHighlight: true,
-                      rowHeight,
-                      rowWidth,
-                      columns: [
-                        {
-                          size: { basis: 350 },
-                          field: 'record_id',
-                          headerRenderer: () => h(Sortable, { sort, field: 'record_id', onSort: setSort }, ['Sample ID']),
-                          cellRenderer: ({ rowIndex }) => {
-                            const engineId = paginatedPreviousRuns[rowIndex].engine_id;
-
-                            // Engine id may be undefined if CBAS failed to submit to Cromwell
-                            if (engineId === undefined) return h(TextCell, [paginatedPreviousRuns[rowIndex].record_id]);
-
-                            return div({ style: { width: '100%', textAlign: 'left' } }, [
-                              h(
-                                Link,
-                                {
-                                  href: Nav.getLink('workspace-workflows-app-run-details', {
-                                    namespace,
-                                    name,
-                                    submissionId,
-                                    workflowId: engineId,
-                                  }),
-                                  style: { fontWeight: 'bold' },
-                                },
-                                [paginatedPreviousRuns[rowIndex].record_id]
-                              ),
-                            ]);
-                          },
-                        },
-                        {
-                          size: { basis: 600, grow: 0 },
-                          field: 'state',
-                          headerRenderer: () => h(Sortable, { sort, field: 'state', onSort: setSort }, ['Status']),
-                          cellRenderer: ({ rowIndex }) => {
-                            const status = state(paginatedPreviousRuns[rowIndex].state, paginatedPreviousRuns[rowIndex].submission_date);
-                            if (errorStates.includes(paginatedPreviousRuns[rowIndex].state)) {
-                              return div({ style: { width: '100%', textAlign: 'center' } }, [
-                                h(Link, { key: 'error link', style: { fontWeight: 'bold' }, onClick: () => setViewErrorsId(rowIndex) }, [
-                                  makeStatusLine((style) => status.icon(style), status.label(paginatedPreviousRuns[rowIndex].state), {
-                                    textAlign: 'center',
-                                  }),
-                                ]),
-                              ]);
-                            }
-                            return h(TextCell, { style: { fontWeight: 'bold' } }, [
-                              makeStatusLine((style) => status.icon(style), status.label(paginatedPreviousRuns[rowIndex].state), {
-                                textAlign: 'center',
-                              }),
-                            ]);
-                          },
-                        },
-                        {
-                          size: { basis: 500, grow: 0 },
-                          field: 'duration',
-                          headerRenderer: () => h(Sortable, { sort, field: 'duration', onSort: setSort }, ['Duration']),
-                          cellRenderer: ({ rowIndex }) => {
-                            return h(TextCell, [customFormatDuration(paginatedPreviousRuns[rowIndex].duration)]);
-                          },
-                        },
-                        {
-                          size: { basis: 400, grow: 0 },
-                          field: 'workflowId',
-                          headerRenderer: () => h(Sortable, { sort, field: 'workflowId', onSort: setSort }, ['Workflow ID']),
-                          cellRenderer: ({ rowIndex }) => {
-                            const engineId = paginatedPreviousRuns[rowIndex].engine_id;
-                            if (engineId !== undefined) {
-                              return h(TextCell, [
-                                span({ style: { marginRight: '0.5rem' } }, [engineId]),
-                                span({}, [h(ClipboardButton, { text: engineId, 'aria-label': 'Copy workflow id' })]),
-                              ]);
-                            }
-                          },
-                        },
-                      ],
-                    }),
-                ]),
-              ]
-            ),
-            !_.isEmpty(sortedPreviousRuns) &&
-              div({ style: { bottom: 0, position: 'absolute', marginBottom: '1.5rem', right: '4rem' } }, [
-                paginator({
-                  filteredDataLength: sortedPreviousRuns.length,
-                  unfilteredDataLength: sortedPreviousRuns.length,
-                  pageNumber,
-                  setPageNumber,
-                  itemsPerPage,
-                  setItemsPerPage: (v) => {
-                    setPageNumber(1);
-                    setItemsPerPage(v);
-                  },
-                  itemsPerPageOptions: [10, 25, 50, 100],
-                }),
-              ]),
-            viewErrorsId !== undefined &&
-              h(
-                Modal,
-                {
-                  title: 'Error Messages',
-                  width: 600,
-                  onDismiss: () => setViewErrorsId(undefined),
-                  showCancel: false,
-                  okButton: h(
-                    ButtonPrimary,
-                    {
-                      disabled: false,
-                      onClick: () => setViewErrorsId(undefined),
-                    },
-                    ['OK']
-                  ),
-                },
-                [
-                  h(
-                    TextCell,
-                    {
-                      style: { textAlign: 'center', whiteSpace: 'pre-wrap', wordBreak: 'break-word', margin: '3rem', marginBottom: '1rem' },
-                    },
-                    [paginatedPreviousRuns[viewErrorsId]?.error_messages]
-                  ),
-                ]
-              ),
-          ]
-        ),
+        div([h(FilterableWorkflowTable, { runsData, runsFullyUpdated, namespace, submissionId, workspaceName: name })]),
       ]);
 };
